revamped the visualization, updated docs
