+++ .gitignore
@@ -54,3 +54,6 @@ coverage.xml
 docs/_build/
 
 .DS_Store
+
+# Generated visualization files
+examples/*.html
