+++ .readthedocs.yaml
@@ -25,3 +25,4 @@ formats:
 python:
   install:
     - requirements: requirements.txt
+    - requirements: docs/requirements.txt
