+++ docs/conf.py
@@ -23,7 +23,7 @@ copyright = "2019, Tim Fiola"
 author = "Tim Fiola"
 
 # The full version, including alpha/beta/rc tags
-release = "4.0.0"
+release = "5.0.0"
 
 # index file
 master_doc = "index"
