+++ docs/examples.rst
@@ -12,7 +12,7 @@ The ``lsp_practice_code.py`` script demos how auto-bandwidth RSVP LSPs react to
 * adding traffic
 * adding additional LSPs
 
-The ``network_modeling_client_code_examples_read_from_dict.py`` file demo2 the following:
+The ``network_modeling_client_code_examples_read_from_dict.py`` file demos the following:
 
 * loading a network topology from a list of info (instead of a model file)
 * addition of new circuit and node to the network
@@ -22,8 +22,19 @@ The ``network_modeling_client_code_examples_read_from_dict.py`` file demo2 the f
 * demand path changes before/after a link failure
 * adding traffic
 
-Demo Script With Visualization
-------------------------------
+Interactive Visualization Example
+----------------------------------
 
-There is also a visualization script, which showcases a current beta feature.
-See the `visualization <visualization.rst>`_ docs page for more details.
+The ``interactive_visualization_example.py`` script demonstrates the interactive
+visualization feature.  It loads a model, runs the simulation, and opens a
+browser-based visualization::
+
+    from pyNTM import Model
+
+    model = Model.load_model_file('sample_network_model_file.csv')
+    model.update_simulation()
+    model.visualize()
+
+You can also call ``model.visualize('output.html')`` to save to a specific file.
+
+See the `visualization <visualization.rst>`_ docs page for full details on visualization features.
