+++ pyNTM/__init__.py
@@ -8,12 +8,14 @@ from .exceptions import ModelException  # noqa: F401
 from .node import Node  # noqa: F401
 from .srlg import SRLG  # noqa: F401
 from .utilities import *  # noqa: F401,F403
-from .model import FlexModel  # noqa: F401
-from .flex_model import Parallel_Link_Model  # noqa: F401
+from .model import Model  # noqa: F401
 from .master_model import _MasterModel  # noqa: F401
 from .interface import Interface  # noqa: F401
 from .rsvp import RSVP_LSP  # noqa: F401
+
+# Backward compatibility aliases
+from .flex_model import FlexModel  # noqa: F401
+from .flex_model import Parallel_Link_Model  # noqa: F401
 from .performance_model import PerformanceModel  # noqa: F401
-from .performance_model import Model  # noqa: F401
 
 # from .weathermap import WeatherMap  # noqa: F401
