+++ pyNTM/performance_model.py
@@ -1,19 +1,12 @@
 """
 Backward compatibility shim.
 
-PerformanceModel is now an alias for FlexModel, defined in pyNTM.model.
-This module re-exports it so that existing code using
-``from pyNTM.performance_model import PerformanceModel`` continues to work.
+PerformanceModel is now a subclass alias for Model, defined in pyNTM.model.
 """
 
-from .model import FlexModel
+from .model import Model
 
 
-class PerformanceModel(FlexModel):
-    """Legacy alias for FlexModel (unified model), kept for backward compatibility."""
-    pass
-
-
-class Model(PerformanceModel):
-    """Legacy alias for backward compatibility with pyNTM <= 1.6."""
+class PerformanceModel(Model):
+    """Backward-compatible alias for Model."""
     pass
