+++ requirements.txt
@@ -1,5 +1,1 @@
 networkx==3.1
-jinja2
-dash
-dash-cytoscape
-sphinx-rtd-theme
